+++ src/clients/catalog.ts
@@ -8,9 +8,11 @@ import { JanusClient } from '@vtex/api'
 import { checkSellerInformation } from '../utils/seller'
 import type { AuthMethod } from '../typings/tokens'
 import type {
-  GetBrandResponse,
   GetSkuResponse,
+  GetSkuContextResponse,
   Seller,
+  Category,
+  GetBrandResponse,
 } from '../typings/catalog'
 import { getRequestConfig } from '../utils/request'
 
@@ -20,14 +22,16 @@ const baseURLLegacy = '/api/catalog_system'
 const routes = {
   productsAndSkus: (from: number, to: number) =>
     `${baseURLLegacy}/pvt/products/GetProductAndSkuIds?_from=${from}&_to=${to}`,
+  getCategoryById: (categoryId: string) =>
+    `${baseURL}/pvt/category/${categoryId}`,
   getSkuById: (skuId: string) => `${baseURL}/pvt/stockkeepingunit/${skuId}`,
+  getSkuContext: (skuId: string) =>
+    `${baseURLLegacy}/pvt/sku/stockkeepingunitbyid/${skuId}`,
   changeNotification: (sellerId: string, skuId: string) =>
     `${baseURLLegacy}/pvt/skuseller/changenotification/${sellerId}/${skuId}`,
   seller: (sellerId: string) => `${baseURLLegacy}/pvt/seller/${sellerId}`,
   sellerList: `${baseURLLegacy}/pvt/seller/list`,
   getBrandById: (brandId: string) => `${baseURLLegacy}/pvt/brand/${brandId}`,
-  getCategoryById: (categoryId: number) =>
-    `${baseURL}/pvt/category/${categoryId}`,
 }
 
 export class Catalog extends JanusClient {
@@ -35,6 +39,19 @@ export class Catalog extends JanusClient {
     super(ctx, {
       ...options,
     })
+  }
+
+  public getCategoryById(
+    categoryId: string,
+    authMethod: AuthMethod = 'AUTH_TOKEN',
+    tracingConfig?: RequestTracingConfig
+  ): Promise<Category | undefined> {
+    const metric = 'catalog-getCategoryMetric'
+
+    return this.http.get<Category>(
+      routes.getCategoryById(categoryId),
+      getRequestConfig(this.context, authMethod, metric, tracingConfig)
+    )
   }
 
   // eslint-disable-next-line max-params
@@ -61,6 +78,19 @@ export class Catalog extends JanusClient {
 
     return this.http.get<GetSkuResponse>(
       routes.getSkuById(skuId),
+      getRequestConfig(this.context, authMethod, metric, tracingConfig)
+    )
+  }
+
+  public getSkuContext(
+    skuId: string,
+    authMethod: AuthMethod = 'AUTH_TOKEN',
+    tracingConfig?: RequestTracingConfig
+  ) {
+    const metric = 'catalog-getSkuContextMetric'
+
+    return this.http.get<GetSkuContextResponse>(
+      routes.getSkuContext(skuId),
       getRequestConfig(this.context, authMethod, metric, tracingConfig)
     )
   }